# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Movie Marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type MovieQuality = typeof MovieQuality[keyof typeof MovieQuality];


export const MovieQuality = {
  '720p': '720p',
  '1080p': '1080p',
  '4K': '4K',
} as const;

export interface Movie {
  id: string;
  title: string;
  description: string;
  posterUrl: string;
  /** @nullable */
  bannerUrl?: string | null;
  /** @nullable */
  youtubeTrailerId?: string | null;
  genre: string[];
  duration: string;
  quality: MovieQuality;
  fileSize: string;
  price: number;
  featured: boolean;
  /** @nullable */
  rating?: number | null;
  year: number;
  /** @nullable */
  telegramFileId?: string | null;
  createdAt: string;
}

export type MovieInputQuality = typeof MovieInputQuality[keyof typeof MovieInputQuality];


export const MovieInputQuality = {
  '720p': '720p',
  '1080p': '1080p',
  '4K': '4K',
} as const;

export interface MovieInput {
  title: string;
  description: string;
  posterUrl: string;
  /** @nullable */
  bannerUrl?: string | null;
  /** @nullable */
  youtubeTrailerId?: string | null;
  genre: string[];
  duration: string;
  quality: MovieInputQuality;
  fileSize: string;
  price: number;
  featured?: boolean;
  /** @nullable */
  rating?: number | null;
  year: number;
  /** @nullable */
  telegramFileId?: string | null;
}

export interface MovieListResponse {
  movies: Movie[];
  total: number;
  page: number;
  totalPages: number;
}

export type MovieStatsByGenre = {[key: string]: number};

export type MovieStatsByQuality = {[key: string]: number};

export interface MovieStats {
  total: number;
  byGenre: MovieStatsByGenre;
  byQuality: MovieStatsByQuality;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  payment_initiated: 'payment_initiated',
  payment_confirmed: 'payment_confirmed',
  delivering: 'delivering',
  delivered: 'delivered',
  failed: 'failed',
} as const;

export type OrderPaymentStatus = typeof OrderPaymentStatus[keyof typeof OrderPaymentStatus];


export const OrderPaymentStatus = {
  pending: 'pending',
  initiated: 'initiated',
  confirmed: 'confirmed',
  failed: 'failed',
} as const;

export interface Order {
  id: string;
  movieId: string;
  movieTitle: string;
  moviePosterUrl: string;
  telegramUsername: string;
  phone: string;
  amount: number;
  status: OrderStatus;
  paymentStatus: OrderPaymentStatus;
  /** @nullable */
  checkoutRequestId?: string | null;
  /** @nullable */
  merchantRequestId?: string | null;
  /** @nullable */
  mpesaReceiptNumber?: string | null;
  /** @nullable */
  deliveredAt?: string | null;
  /** @nullable */
  failureReason?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface OrderInput {
  movieId: string;
  telegramUsername: string;
  phone: string;
}

export type MpesaCallbackInputBody = { [key: string]: unknown };

export interface MpesaCallbackInput {
  Body: MpesaCallbackInputBody;
}

export interface PaymentStatus {
  status: string;
  /** @nullable */
  resultCode?: string | null;
  /** @nullable */
  resultDesc?: string | null;
}

export type ListMoviesParams = {
genre?: string;
quality?: string;
search?: string;
page?: number;
limit?: number;
};