# lib/api-zod/src/generated/types/movieInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Movie Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { MovieInputQuality } from './movieInputQuality';

export interface MovieInput {
  title: string;
  description: string;
  posterUrl: string;
  /** @nullable */
  bannerUrl?: string | null;
  /** @nullable */
  youtubeTrailerId?: string | null;
  genre: string[];
  duration: string;
  quality: MovieInputQuality;
  fileSize: string;
  price: number;
  featured?: boolean;
  /** @nullable */
  rating?: number | null;
  year: number;
  /** @nullable */
  telegramFileId?: string | null;
}