# lib/api-zod/src/generated/types/movieListResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Movie Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { Movie } from './movie';

export interface MovieListResponse {
  movies: Movie[];
  total: number;
  page: number;
  totalPages: number;
}