# lib/api-zod/src/generated/types/mpesaCallbackInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Movie Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { MpesaCallbackInputBody } from './mpesaCallbackInputBody';

export interface MpesaCallbackInput {
  Body: MpesaCallbackInputBody;
}