# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Movie Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderPaymentStatus } from './orderPaymentStatus';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: string;
  movieId: string;
  movieTitle: string;
  moviePosterUrl: string;
  telegramUsername: string;
  phone: string;
  amount: number;
  status: OrderStatus;
  paymentStatus: OrderPaymentStatus;
  /** @nullable */
  checkoutRequestId?: string | null;
  /** @nullable */
  merchantRequestId?: string | null;
  /** @nullable */
  mpesaReceiptNumber?: string | null;
  /** @nullable */
  deliveredAt?: string | null;
  /** @nullable */
  failureReason?: string | null;
  createdAt: string;
  updatedAt: string;
}