# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Movie Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderInput {
  movieId: string;
  telegramUsername: string;
  phone: string;
}